+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.1.4
+###Added
+* Pdf generator
+
+
+## 0.1.3
+###Added
+* Asset request
+* Requested assets
+
 ## 0.1.2
 ### Added
 * Organization assets